requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
torch==2.1.0
sentence-transformers==2.2.2
transformers==4.35.0
networkx==3.2.1

# Google Gemini AI
google-generativeai==0.3.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0

# Data Processing
numpy==1.24.3
